+++ rastervision/registry.py
@@ -153,9 +153,9 @@ class Registry:
         }
 
         self.experiment_runners = {
-            rv.OLD_LOCAL: rv.runner.LocalExperimentRunner,
+            rv.OLD_LOCAL: rv.runner.OldLocalExperimentRunner,
             rv.AWS_BATCH: rv.runner.AwsBatchExperimentRunner,
-            rv.LOCAL: rv.runner.NewLocalExperimentRunner
+            rv.LOCAL: rv.runner.LocalExperimentRunner
         }
 
         self.filesystems = [
